Network.Wai.Middleware.Servant.Options

Description

A middleware to respond to Options requests for a servant app very helpful when trying to deal with pre-flight CORS requests.

Documentation

provideOptions :: (GenerateList NoContent (Foreign NoContent api), HasForeign NoTypes NoContent api) => Proxy api -> Middleware